+++ lib/common/constants.dart
@@ -19,6 +19,7 @@ class ZOColors {
   static const successLight = Color.fromRGBO(231, 240, 204, 1);
   static const onSuccess = Color.fromRGBO(255, 255, 255, 1);
   static const amberTransparent = Color.fromRGBO(255, 182, 0, 0.4);
+  static const assistChipSelectedBackground = Color.fromRGBO(73, 69, 79, 0.12);
 }
 
 class ZOStrings {